Refactor SMS sender related extensions
