+++ features/admin.extensions.v1/components/identity-providers/constants/index.ts
@@ -17,4 +17,3 @@
  */
 
 export * from "./swe-constants";
-export * from "./sms-otp-constants";
