+++ features/admin.extensions.v1/components/identity-providers/models/index.ts
@@ -17,4 +17,4 @@
  */
 
 export * from "./swe";
-export * from "./sms-otp";
+export * from "@wso2is/admin.sms-providers.v1/models/sms-otp";
